+++ packages/ui/src/calendar/test/range-calendar.test.tsx
@@ -117,8 +117,8 @@ describe( 'RangeCalendar', () => {
 			render( <RangeCalendar /> );
 
 			expect(
-				screen.getByRole( 'group', {
-					name: 'Date range calendar',
+				screen.getByRole( 'application', {
+					name: 'Date range calendar, May 2025',
 				} )
 			).toBeVisible();
 
@@ -138,6 +138,11 @@ describe( 'RangeCalendar', () => {
 
 		it( 'should show multiple months at once via the `numberOfMonths` prop', () => {
 			render( <RangeCalendar numberOfMonths={ 2 } /> );
+			expect(
+				screen.getByRole( 'application', {
+					name: 'Date range calendar, May 2025 and June 2025',
+				} )
+			).toBeVisible();
 
 			const grids = screen.getAllByRole( 'grid' );
 			expect( grids ).toHaveLength( 2 );
@@ -1505,8 +1510,8 @@ describe( 'RangeCalendar', () => {
 
 			// Check computed writing direction
 			expect(
-				screen.getByRole( 'group', {
-					name: 'Date range calendar',
+				screen.getByRole( 'application', {
+					name: /^Date range calendar,/,
 				} )
 			).toHaveAttribute( 'dir', 'rtl' );